requirements.txt
appdirs>=1.4.0
autopep8>=1.2.4
coreapi>=2.2.4
coreschema>=0.0.4
opencv-python>=3.1.0.2
dj-database-url>=0.4.2
Django>=1.9.6
django-registration-redux>=1.3
django-rest-swagger>=2.1.1
djangorestframework>=3.5.4
hellowebapp-deploy>=1.0.2
httpie>=0.9.9
itypes>=1.1.0
Jinja2>=2.9.5
libsass>=0.12.3
MarkupSafe>=0.23
matplotlib>=1.5.1
numpy>=1.11.2
olefile>=0.44
openapi-codec>=1.3.1
packaging>=16.8
pep8>=1.7.0
Pillow>=4.0.0
Pygments>=2.2.0
pyparsing>=2.1.10
requests>=2.13.0
simplejson>=3.10.0
six>=1.10.0
scikit-image>=0.12.3
scikit-learn>=0.18.1
scipy>=0.18.1
uritemplate>=3.0.0
waitress>=1.0.2
whitenoise>=3.3.0
